grading-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
easyocr==1.7.1
paddlepaddle==2.6.1
paddleocr==2.7.3
Pillow==10.4.0
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.0
supabase==2.7.2
python-dotenv==1.0.1
pdfplumber==0.11.4
httpx==0.27.0
apscheduler==3.10.4
numpy==1.26.4
